code review - name changes
