+++ securityconfig/elasticsearch.yml.example
@@ -221,7 +221,7 @@ opendistro_security.audit.type: internal_elasticsearch
 # opendistro_security.protected_indices.indices: ['.opendistro-alerting-config', '.opendistro-ism-*']
 
 
-opendistro_security.evaluator: 'com.amazon.opendistroforelasticsearch.security.privileges.RangerPrivilegesEvaluator'
+opendistro_security.privilegesEvaluator: 'com.amazon.opendistroforelasticsearch.security.privileges.RangerPrivilegesEvaluator'
 opendistro_security.ranger.app_id: 'rafd002_elasticsearch'
 opendistro_security.kerberos.acceptor_principal: 'HTTP/rafd002-slv-01.cloud.in.guavus.com@GVS.GGN
 opendistro_security.kerberos.acceptor_keytab_filepath: 'spnego.service.keytab'
