+++ src/main/java/com/amazon/opendistroforelasticsearch/security/configuration/OpenDistroSecurityIndexSearcherWrapper.java
@@ -34,7 +34,6 @@ import java.io.IOException;
 import java.util.Collection;
 import java.util.Set;
 
-import com.amazon.opendistroforelasticsearch.security.privileges.Evaluator;
 import com.amazon.opendistroforelasticsearch.security.privileges.PrivilegesEvaluator;
 import com.amazon.opendistroforelasticsearch.security.support.WildcardMatcher;
 import org.apache.logging.log4j.LogManager;
@@ -60,17 +59,17 @@ public class OpenDistroSecurityIndexSearcherWrapper extends IndexSearcherWrapper
     protected final Index index;
     protected final String opendistrosecurityIndex;
     private final AdminDNs adminDns;
-    private final Evaluator evaluator;
+    private final PrivilegesEvaluator privilegesEvaluator;
     private final Collection<String> indexPatterns;
     private final Collection<String> allowedRoles;
     private final Boolean protectedIndexEnabled;
 
     //constructor is called per index, so avoid costly operations here
-    public OpenDistroSecurityIndexSearcherWrapper(final IndexService indexService, final Settings settings, final AdminDNs adminDNs, final Evaluator evaluator) {
+    public OpenDistroSecurityIndexSearcherWrapper(final IndexService indexService, final Settings settings, final AdminDNs adminDNs, final PrivilegesEvaluator privilegesEvaluator) {
         index = indexService.index();
         threadContext = indexService.getThreadPool().getThreadContext();
         this.opendistrosecurityIndex = settings.get(ConfigConstants.OPENDISTRO_SECURITY_CONFIG_INDEX_NAME, ConfigConstants.OPENDISTRO_SECURITY_DEFAULT_CONFIG_INDEX);
-        this.evaluator = evaluator;
+        this.privilegesEvaluator = privilegesEvaluator;
         this.adminDns = adminDNs;
         this.indexPatterns = settings.getAsList(ConfigConstants.OPENDISTRO_SECURITY_PROTECTED_INDICES_KEY);
         this.allowedRoles = settings.getAsList(ConfigConstants.OPENDISTRO_SECURITY_PROTECTED_INDICES_ROLES_KEY);
@@ -132,7 +131,7 @@ public class OpenDistroSecurityIndexSearcherWrapper extends IndexSearcherWrapper
         if (caller == null || user == null) {
             return false;
         }
-        final Set<String> securityRoles = evaluator.mapRoles(user, caller);
+        final Set<String> securityRoles = privilegesEvaluator.mapRoles(user, caller);
         if (WildcardMatcher.matchAny(allowedRoles, securityRoles)) {
             return true;
         }
