+++ src/main/java/com/amazon/opendistroforelasticsearch/security/dlic/rest/api/GetConfigurationApiAction.java
@@ -19,7 +19,7 @@ package com.amazon.opendistroforelasticsearch.security.dlic.rest.api;
 import java.nio.file.Path;
 import java.util.Set;
 
-import com.amazon.opendistroforelasticsearch.security.privileges.Evaluator;
+import com.amazon.opendistroforelasticsearch.security.privileges.PrivilegesEvaluator;
 import org.elasticsearch.client.Client;
 import org.elasticsearch.cluster.service.ClusterService;
 import org.elasticsearch.common.bytes.BytesReference;
@@ -39,7 +39,6 @@ import com.amazon.opendistroforelasticsearch.security.configuration.AdminDNs;
 import com.amazon.opendistroforelasticsearch.security.configuration.IndexBaseConfigurationRepository;
 import com.amazon.opendistroforelasticsearch.security.dlic.rest.validation.AbstractConfigurationValidator;
 import com.amazon.opendistroforelasticsearch.security.dlic.rest.validation.NoOpValidator;
-import com.amazon.opendistroforelasticsearch.security.privileges.PrivilegesEvaluator;
 import com.amazon.opendistroforelasticsearch.security.ssl.transport.PrincipalExtractor;
 import com.amazon.opendistroforelasticsearch.security.support.ConfigConstants;
 import com.google.common.base.Joiner;
@@ -53,8 +52,8 @@ public class GetConfigurationApiAction extends AbstractApiAction {
 	@Inject
 	public GetConfigurationApiAction(final Settings settings, final Path configPath, final RestController controller, final Client client,
 									 final AdminDNs adminDNs, final IndexBaseConfigurationRepository cl, final ClusterService cs,
-									 final PrincipalExtractor principalExtractor, final Evaluator evaluator, ThreadPool threadPool, AuditLog auditLog) {
-		super(settings, configPath, controller, client, adminDNs, cl, cs, principalExtractor, evaluator, threadPool, auditLog);
+									 final PrincipalExtractor principalExtractor, final PrivilegesEvaluator privilegesEvaluator, ThreadPool threadPool, AuditLog auditLog) {
+		super(settings, configPath, controller, client, adminDNs, cl, cs, principalExtractor, privilegesEvaluator, threadPool, auditLog);
 	}
 
 	@Override
