+++ src/main/java/com/amazon/opendistroforelasticsearch/security/privileges/PrivilegesEvaluatorResponse.java
@@ -1,99 +1,21 @@
-/*
- * Copyright 2015-2018 _floragunn_ GmbH
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/*
- * Portions Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License").
- * You may not use this file except in compliance with the License.
- * A copy of the License is located at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * or in the "license" file accompanying this file. This file is distributed
- * on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
- * express or implied. See the License for the specific language governing
- * permissions and limitations under the License.
- */
-
 package com.amazon.opendistroforelasticsearch.security.privileges;
 
-import java.util.HashSet;
 import java.util.Map;
 import java.util.Set;
 
-public class PrivilegesEvaluatorResponse implements EvaluatorResponse {
-    boolean allowed = false;
-    Set<String> missingPrivileges = new HashSet<String>();
-    Map<String,Set<String>> allowedFlsFields;
-    Map<String,Set<String>> maskedFields;
-    Map<String,Set<String>> queries;
-    PrivilegesEvaluatorResponseState state = PrivilegesEvaluatorResponseState.PENDING;
+/**
+ * @author Divyansh Jain
+ */
 
-    @Override
-    public boolean isAllowed() {
-        return allowed;
-    }
+public interface PrivilegesEvaluatorResponse {
 
-    @Override
-    public Set<String> getMissingPrivileges() {
-        return new HashSet<String>(missingPrivileges);
-    }
+    boolean isAllowed();
 
-    @Override
-    public Map<String,Set<String>> getAllowedFlsFields() {
-        return allowedFlsFields;
-    }
+    Map<String, Set<String>> getAllowedFlsFields();
 
-    @Override
-    public Map<String,Set<String>> getMaskedFields() {
-        return maskedFields;
-    }
+    Map<String, Set<String>> getMaskedFields();
 
-    @Override
-    public Map<String,Set<String>> getQueries() {
-        return queries;
-    }
+    Map<String, Set<String>> getQueries();
 
-    public PrivilegesEvaluatorResponse markComplete() {
-        this.state = PrivilegesEvaluatorResponseState.COMPLETE;
-        return this;
-    }
-
-    public PrivilegesEvaluatorResponse markPending() {
-        this.state = PrivilegesEvaluatorResponseState.PENDING;
-        return this;
-    }
-
-    public boolean isComplete() {
-        return this.state == PrivilegesEvaluatorResponseState.COMPLETE;
-    }
-
-    public boolean isPending() {
-        return this.state == PrivilegesEvaluatorResponseState.PENDING;
-    }
-
-    @Override
-    public String toString() {
-        return "PrivEvalResponse [allowed=" + allowed + ", missingPrivileges=" + missingPrivileges
-                + ", allowedFlsFields=" + allowedFlsFields + ", maskedFields=" + maskedFields + ", queries=" + queries + "]";
-    }
-
-    public static enum PrivilegesEvaluatorResponseState {
-        PENDING,
-        COMPLETE;
-    }
-
+    Set<String> getMissingPrivileges();
 }
