+++ src/test/java/com/amazon/opendistroforelasticsearch/security/dlic/rest/api/RestApiPrivilegesEvaluatorTest.java
@@ -16,7 +16,7 @@
 package com.amazon.opendistroforelasticsearch.security.dlic.rest.api;
 
 import com.amazon.opendistroforelasticsearch.security.configuration.AdminDNs;
-import com.amazon.opendistroforelasticsearch.security.privileges.PrivilegesEvaluator;
+import com.amazon.opendistroforelasticsearch.security.privileges.OpenDistroPrivilegesEvaluator;
 import com.amazon.opendistroforelasticsearch.security.ssl.transport.PrincipalExtractor;
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.rest.RestRequest;
@@ -39,7 +39,7 @@ public class RestApiPrivilegesEvaluatorTest {
     public void setUp() {
         this.privilegesEvaluator = new RestApiPrivilegesEvaluator(Settings.EMPTY,
                 mock(AdminDNs.class),
-                mock(PrivilegesEvaluator.class),
+                mock(OpenDistroPrivilegesEvaluator.class),
                 mock(PrincipalExtractor.class),
                 mock(Path.class),
                 mock(ThreadPool.class));
